fix bugs in edit mode

+++ src/cmps/movieDetails/MovieDetails.js
@@ -3,25 +3,29 @@ import './MovieDetails.css';
 import MovieService from '../../services/MovieService';
 import { Route, Redirect } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-
+import swal from "sweetalert";
 
 export default class MovieDetails extends Component {
     constructor(props, context) {
         super(props, context);
         this.state = {
-        movieId: this.props.match.params.movieId,
-        movie: {},
-        back: false,
-        edit: false
+            movieId: this.props.match.params.movieId,
+            movie: {},
+            // movie: this.props.location,
+            back: false,
+            edit: false
         }
     }
     componentDidMount() {
-        // console.log('id', this.state.MovieId);
-        MovieService.getMovieById(this.state.movieId)
-        .then(data => {
-            console.log('movie$$', data);
-            this.setState({movie: data});
-        });
+        // MovieService.getMovieById(this.state.movieId)
+        // .then(data => {
+        //     // console.log('movie$$', data);
+        //     this.setState({movie: data});
+        // });
+        console.log('id', this.state.MovieId);
+        var currMovie = MovieService.searchMovieById(this.state.movieId)
+            console.log('movie$$', currMovie);
+        this.setState({movie: currMovie});
     }
     editMovie = () => {
         // console.log('movies: ', MovieService.getUpdatedMovies());
@@ -32,8 +36,26 @@ export default class MovieDetails extends Component {
         // console.log('back', this.state.back);
     }
     delete = () => {
-        MovieService.deleteMovie(this.state.movieId);
-        this.setState({back: !this.state.back})
+        swal({
+            title: "Are you sure you want to delete this movie?",
+            icon: "warning",
+            buttons: ["Cancel", "Delete"],
+            dangerMode: true,
+            className: "swal-warning"
+        }).then(willDelete => {
+            if (willDelete) {
+            MovieService.deleteMovie(this.state.movieId)
+                .then(() => {
+                    this.setState({back: !this.state.back})
+                    swal("Your movie has been deleted!", {
+                    icon: "success",
+                    timer: 2000,
+                    className: "swal-text",
+                    button: false
+                    });
+                });
+            } else swal.close();
+        });
     }
     render() {
         if (this.state.back) {
@@ -43,24 +65,26 @@ export default class MovieDetails extends Component {
             return <Redirect to={{pathname: `/movie/edit/${this.props.match.params.movieId}`, refreshMovies: this.props.location.refreshMovies}} />
         }
         const movie = this.state.movie;
+        var imgSrc = movie.Poster;
+        if (imgSrc === null || imgSrc === 'N/A') imgSrc = 'img/movie3.png'
         return (
             <div className="movie-details-container flex column">
                 <h2 className="title-details">{movie.Title}</h2>
                 <div className="movie-details flex column">
                     <div className="details-container flex">
-                        <img src={movie.Poster} alt=""/>
+                        <img src={imgSrc || 'img/movie3.png'} alt=""/>
                         <div className="movie-text flex column">
                         {/* <div class="movie-info flex column"> */}
                             <p className="director"><span>Directed by: </span>{movie.Director}</p>
-                            <p className="actors"><span>Actors: </span>{movie.Actors}</p>
-                            <p className="rating"><span>IMDb rating: </span>{movie.imdbRating}/10</p>
+                            <p className="actors"><span>Actors: </span>{movie.Actors? movie.Actors: 'Gal Gadot' }</p>
+                            <p className="rating"><span>IMDb rating: </span>{movie.imdbRating? movie.imdbRating: '2.5'}/10</p>
                         {/* </div> */}
                             <div className="year-container flex">
                                 <p className="year">{movie.Year},&nbsp;</p>
                                 <p className="runtime"> {movie.Runtime},&nbsp;</p>
-                                <p className="language"> {movie.Language}</p>
+                                <p className="language"> {movie.Language? movie.Language: 'Jibrish'}</p>
                             </div>
-                            <p className="plot">{movie.Plot}</p>
+                            <p className="plot">{movie.Plot? movie.Plot: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident.'}</p>
                         </div>
                     </div>
                     <div className="btns flex">
